project/requirements.txt
attrs==23.1.0
certifi==2023.7.22
config==0.5.1
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
PySocks==1.7.1
python-decouple==3.8
selenium==4.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
